+++ idea/src/org/jetbrains/kotlin/idea/intentions/ImplementAbstractMemberIntention.kt
@@ -41,13 +41,12 @@ import org.jetbrains.kotlin.idea.caches.resolve.getJavaClassDescriptor
 import org.jetbrains.kotlin.idea.caches.resolve.resolveToDescriptorIfAny
 import org.jetbrains.kotlin.idea.core.overrideImplement.OverrideImplementMembersHandler
 import org.jetbrains.kotlin.idea.core.overrideImplement.OverrideMemberChooserObject
-import org.jetbrains.kotlin.idea.refactoring.isInterfaceClass
+import org.jetbrains.kotlin.idea.refactoring.isAbstract
 import org.jetbrains.kotlin.idea.runSynchronouslyWithProgress
 import org.jetbrains.kotlin.idea.search.declarationsSearch.HierarchySearchRequest
 import org.jetbrains.kotlin.idea.search.declarationsSearch.searchInheritors
 import org.jetbrains.kotlin.idea.util.application.executeCommand
 import org.jetbrains.kotlin.idea.util.application.runWriteAction
-import org.jetbrains.kotlin.lexer.KtTokens
 import org.jetbrains.kotlin.load.java.descriptors.JavaClassDescriptor
 import org.jetbrains.kotlin.psi.*
 import org.jetbrains.kotlin.psi.psiUtil.containingClassOrObject
@@ -62,16 +61,6 @@ abstract class ImplementAbstractMemberIntentionBase :
         SelfTargetingRangeIntention<KtNamedDeclaration>(KtNamedDeclaration::class.java, "", "Implement abstract member") {
     companion object {
         private val LOG = Logger.getInstance("#${ImplementAbstractMemberIntentionBase::class.java.canonicalName}")
-    }
-
-    private fun isAbstract(element: KtNamedDeclaration): Boolean {
-        if (element.hasModifier(KtTokens.ABSTRACT_KEYWORD)) return true
-        if (!(element.containingClassOrObject?.isInterfaceClass() ?: false)) return false
-        return when (element) {
-            is KtProperty -> element.initializer == null && element.delegate == null && element.accessors.isEmpty()
-            is KtNamedFunction -> !element.hasBody()
-            else -> false
-        }
     }
 
     protected fun findExistingImplementation(
@@ -117,7 +106,7 @@ abstract class ImplementAbstractMemberIntentionBase :
     protected abstract fun computeText(element: KtNamedDeclaration): String?
 
     override fun applicabilityRange(element: KtNamedDeclaration): TextRange? {
-        if (!isAbstract(element)) return null
+        if (!element.isAbstract()) return null
 
         text = computeText(element) ?: return null