+++ idea/src/org/jetbrains/kotlin/idea/refactoring/pullUp/KotlinPullUpDialog.kt
@@ -77,6 +77,7 @@ class KotlinPullUpDialog(
             if (member.hasModifier(KtTokens.INLINE_KEYWORD) ||
                 member.hasModifier(KtTokens.EXTERNAL_KEYWORD) ||
                 member.hasModifier(KtTokens.LATEINIT_KEYWORD)) return false
+            if (member.isAbstractInInterface(sourceClass)) return false
             if (member.isCompanionMemberOf(sourceClass)) return false
 
             if (!superClass.isInterface()) return true
@@ -87,6 +88,7 @@ class KotlinPullUpDialog(
         override fun isAbstractWhenDisabled(memberInfo: KotlinMemberInfo): Boolean {
             val member = memberInfo.member
             if (member.isCompanionMemberOf(sourceClass)) return false
+            if (member.isAbstractInInterface(sourceClass)) return true
             return ((member is KtProperty || member is KtParameter) && superClass !is PsiClass)
                    || (member is KtNamedFunction && superClass is PsiClass)
         }